migrate java8 to same logic as java11
