+++ src/main/java/com/microsoft/azure/functions/worker/reflect/FunctionClassLoaderProvider.java
@@ -9,8 +9,8 @@ import java.util.concurrent.*;
 
 import com.microsoft.azure.functions.worker.*;
 
-public class EnhancedClassLoaderProvider implements ClassLoaderProvider {
-    public EnhancedClassLoaderProvider() {
+public class FunctionClassLoaderProvider implements ClassLoaderProvider {
+    public FunctionClassLoaderProvider() {
         customerUrls = Collections.newSetFromMap(new ConcurrentHashMap<URL, Boolean>());
         workerUrls = Collections.newSetFromMap(new ConcurrentHashMap<URL, Boolean>());
     }
@@ -28,29 +28,11 @@ public class EnhancedClassLoaderProvider implements ClassLoaderProvider {
                     urlsList.addAll(workerUrls);
                     URL[] urlsForClassLoader = urlsList.toArray(new URL[0]);
                     URLClassLoader loader = new URLClassLoader(urlsForClassLoader);
-                    loadDrivers(loader);
                     classLoaderInstance = loader;
                 }
             }
         }
         return classLoaderInstance;
-    }
-
-    private void loadDrivers(URLClassLoader classLoader) {
-        Thread.currentThread().setContextClassLoader(classLoader);
-        try {
-            ServiceLoader<Driver> loadedDrivers = ServiceLoader.load(Driver.class);
-            Iterator<Driver> driversIterator = loadedDrivers.iterator();
-            try {
-                while (driversIterator.hasNext()) {
-                    driversIterator.next();
-                }
-            } catch (Throwable t) {
-                // Do nothing
-            }
-        } finally {
-            Thread.currentThread().setContextClassLoader(ClassLoader.getSystemClassLoader());
-        }
     }
 
     @Override
